requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
google-genai>=0.3.0
watchdog>=4.0.0
# Optional drag-and-drop support:
tkinterdnd2>=0.3.0
